+++ source/rendering.cpp
@@ -61,8 +61,8 @@ void fill_background(uint16_t color){
    memset32(background, fill_val, SCREEN_WIDTH * SCREEN_HEIGHT / 2);
 }
 
-void restore_background(uint16_t x, uint16_t y, uint16_t w, uint16_t h){
-   for(uint16_t yinc = 0; yinc < h; yinc++){
+void restore_background(int32_t x, int32_t y, int32_t w, int32_t h){
+   for(int32_t yinc = 0; yinc < h; yinc++){
       memcpy16(&vram[x + ((y + yinc) * SCREEN_WIDTH)], &background[x + ((y + yinc) * SCREEN_WIDTH)], w);
    }
 }
@@ -70,10 +70,10 @@ void restore_background(uint16_t x, uint16_t y, uint16_t w, uint16_t h){
 void restore_background(entity& ent){
    if(!ent.dirty.is_dirty)return;
    
-   int16_t clip_x = (int16_t)ent.dirty.x;
-   int16_t clip_y = (int16_t)ent.dirty.y;
-   int16_t width = ent.sprite.w;
-   int16_t height = ent.sprite.h;
+   int32_t clip_x = ent.dirty.x;
+   int32_t clip_y = ent.dirty.y;
+   int32_t width  = ent.sprite.w;
+   int32_t height = ent.sprite.h;
    
    while(clip_x < 0){
       clip_x++;
@@ -97,13 +97,13 @@ void restore_background(entity& ent){
    ent.dirty.is_dirty = false;
 }
 
-void draw_texture(int16_t x, int16_t y, texture& tex){
+void draw_texture(int32_t x, int32_t y, texture& tex){
    if(tex.bitmap == NULL)return;
 
    if(x < 0 || y < 0 || x + tex.w >= SCREEN_WIDTH || y + tex.h >= SCREEN_HEIGHT){
       //if offscreen rendering, use this safe function
-      for(int16_t yinc = 0; yinc < tex.h; yinc++){
-         for(int16_t xinc = 0; xinc < tex.w; xinc++){
+      for(int32_t yinc = 0; yinc < tex.h; yinc++){
+         for(int32_t xinc = 0; xinc < tex.w; xinc++){
             //check "is visible" bit
             uint16_t color = tex.bitmap[xinc + (yinc * tex.w)];
             if(color & 0x8000 && x + xinc > -1 && y + yinc > -1 && x + xinc < SCREEN_WIDTH && y + yinc < SCREEN_HEIGHT){
@@ -114,8 +114,8 @@ void draw_texture(int16_t x, int16_t y, texture& tex){
    }
    else{
       //not on an edge, use this fast function
-      for(int16_t yinc = 0; yinc < tex.h; yinc++){
-         for(int16_t xinc = 0; xinc < tex.w; xinc++){
+      for(int32_t yinc = 0; yinc < tex.h; yinc++){
+         for(int32_t xinc = 0; xinc < tex.w; xinc++){
             //check "is visible" bit
             uint16_t color = tex.bitmap[xinc + (yinc * tex.w)];
             if(color & 0x8000){
@@ -127,10 +127,10 @@ void draw_texture(int16_t x, int16_t y, texture& tex){
 
 }
 
-void draw_texture_background(uint16_t x, uint16_t y, texture& tex){
+void draw_texture_background(int32_t x, int32_t y, texture& tex){
    if(tex.bitmap == NULL)return;
-   for(uint16_t yinc = 0; yinc < tex.h; yinc++){
-      for(uint16_t xinc = 0; xinc < tex.w; xinc++){
+   for(int32_t yinc = 0; yinc < tex.h; yinc++){
+      for(int32_t xinc = 0; xinc < tex.w; xinc++){
          //check "is visible" bit
          uint16_t color = tex.bitmap[xinc + (yinc * tex.w)];
          if(color & 0x8000){
